// types.ts
// 학생 정보
export interface AccountInfo {
    username: string;
    email: string;
    password: string; 
    studentnumber: number;
    //phone: string;
    //lesson: string;

}


// TODO: 이거 이름 못 알아보겠어요ㅠ
// 핵심 역량 정보
export interface CoreCompetencies {
    serlead: number;
    sympcomm: number;
    originality: number;
    pliability: number;
    knowledgequest: number;
    psskills: number;
    globalmind: number;
    asianculture: number;
}